pyproject.toml
[project]
name = "pokemon-agent"
version = "0.1.0"
description = "Pokemon information agent using ADK, MCP and A2A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk==0.3.3",
    "fastmcp==2.11.3",
    "google-adk==1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
]